simplify offset tracker
