+++ connector/src/test/java/io/questdb/kafka/QuestDBSinkConnectorEmbeddedTest.java
@@ -24,7 +24,6 @@ import org.testcontainers.containers.FixedHostPortGenericContainer;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.containers.output.Slf4jLogConsumer;
 import org.testcontainers.containers.wait.strategy.LogMessageWaitStrategy;
-import org.testcontainers.containers.wait.strategy.ShellStrategy;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
 import java.math.BigDecimal;
@@ -66,32 +65,7 @@ public final class QuestDBSinkConnectorEmbeddedTest {
     @AfterAll
     public static void stopContainer() {
         questDBContainer.stop();
-//        deleteFromContainer("questdb");
         Files.rmdir(io.questdb.std.str.Path.getThreadLocal(dbRoot.toAbsolutePath().toString()));
-    }
-
-    private static void deleteFromContainer(String directory) {
-        GenericContainer<?> cleanup = new GenericContainer<>("alpine:3.18.4")
-                .withFileSystemBind(dbRoot.toAbsolutePath().toString(), "/var/lib/delete")
-                .withCommand("ls -l /var/lib/delete/")
-                .withLogConsumer(new Slf4jLogConsumer(LoggerFactory.getLogger("cleanup1")));
-        cleanup.start();
-        cleanup.stop();
-
-        cleanup = new GenericContainer<>("alpine:3.18.4")
-                .withFileSystemBind(dbRoot.toAbsolutePath().toString(), "/var/lib/delete")
-                .withCommand("rm -rf /var/lib/delete/" + directory)
-                .withLogConsumer(new Slf4jLogConsumer(LoggerFactory.getLogger("cleanup2")))
-                .waitingFor(new ShellStrategy().withCommand("rm -rf /var/lib/delete/" + directory));
-        cleanup.start();
-        cleanup.stop();
-
-        cleanup = new GenericContainer<>("alpine:3.18.4")
-                .withFileSystemBind(dbRoot.toAbsolutePath().toString(), "/var/lib/delete")
-                .withCommand("ls -l /var/lib/delete/")
-                .withLogConsumer(new Slf4jLogConsumer(LoggerFactory.getLogger("cleanup3")));
-        cleanup.start();
-        cleanup.stop();
     }
 
     private static String questDBDirectory() {
